Added a function that lets you define a search string with visual select
This function and key mapping lets you define a visual select block,
which will then be piped into the document wide search and replace
command. It pauses to let you fill in the replace string. The function
also can deal with regex charcters in the search string, by escaping
them for the search command.